Fixes from Thomson's PR comments
